update and delete linked to user_id (index - view)

+++ resources/views/movies/index.blade.php
@@ -24,10 +24,12 @@
                     @endif
                     {{ $movie->duration() }}</p>
                 </a> 
-                <div class="mx-3 mb-3 flex justify-between gap-2">
-                    <a href="/film/{{$movie->id}}/modifier" class="text-sm bg-gray-500 px-2 py-1 text-gray-200 w-1/2 text-center">Modifier</a>
-                    <a href="/film/{{$movie->id}}/supprimer" class="text-sm bg-red-500 px-2 py-1 text-gray-200 w-1/2 text-center" onclick="return confirm('Êtes-vous sûr de vouloir supprimer le film {{$movie->title}}')">Supprimer</a>
-                </div>
+                @if (Auth::user() && Auth::user()->id == $movie->user_id) {{-- On affiche modifier/supprimer si on est conencté et qu'on a le film--}}
+                    <div class="mx-3 mb-3 flex justify-between gap-2">
+                        <a href="/film/{{$movie->id}}/modifier" class="text-sm bg-gray-500 px-2 py-1 text-gray-200 w-1/2 text-center">Modifier</a>
+                        <a href="/film/{{$movie->id}}/supprimer" class="text-sm bg-red-500 px-2 py-1 text-gray-200 w-1/2 text-center" onclick='return confirm("Êtes-vous sûr de vouloir supprimer le film {{$movie->title}}")'>Supprimer</a>
+                    </div>
+                @endif
             </div>
         </div>    
     @endforeach
